(#332) Move TextEditLog to separate file
