+++ include/log.h
@@ -25,24 +25,4 @@ public:
     ~Log() = default;
 };
 
-class TextEditLog : public Log {
-private:
-    QTextEdit *console;
-public:
-    virtual void debug(const QString &message) override;
-    virtual void info(const QString &message) override;
-    virtual void warning(const QString &message) override;
-    virtual void error(const QString &message) override;
-    virtual void debugT(const QString &message) override;
-    virtual void infoT(const QString &message) override;
-    virtual void warningT(const QString &message) override;
-    virtual void errorT(const QString &message) override;
-    virtual void write(const QString &message) override;
-    virtual void writeln(const QString &message) override;
-    virtual void setColor(const QString &color) override;
-    virtual void resetColor() override;
-    TextEditLog(Game *game, QTextEdit *logField);
-    ~TextEditLog() = default;
-};
-
 #endif  // INCLUDE_LOG_H_
